Fix comparing paths in getDistanceFromPathInRootNavigator
