+++ src/libs/NavigationUtils.ts
@@ -1,5 +1,7 @@
+import _ from 'lodash';
 import SCREENS from '@src/SCREENS';
-import type {CentralPaneName} from './Navigation/types';
+import getTopmostBottomTabRoute from './Navigation/getTopmostBottomTabRoute';
+import type {CentralPaneName, RootStackParamList, State} from './Navigation/types';
 
 const CENTRAL_PANE_SCREEN_NAMES = new Set([
     SCREENS.SETTINGS.WORKSPACES,
@@ -23,4 +25,13 @@ function isCentralPaneName(screen: string | undefined): screen is CentralPaneNam
     return CENTRAL_PANE_SCREEN_NAMES.has(screen as CentralPaneName);
 }
 
-export default isCentralPaneName;
+const removePolicyIDParamFromState = (state: State<RootStackParamList>) => {
+    const stateCopy = _.cloneDeep(state);
+    const bottomTabRoute = getTopmostBottomTabRoute(stateCopy);
+    if (bottomTabRoute?.params && 'policyID' in bottomTabRoute.params) {
+        delete bottomTabRoute.params.policyID;
+    }
+    return stateCopy;
+};
+
+export {isCentralPaneName, removePolicyIDParamFromState};
